refactor(utils): flatten utils package
Break down utils package to `executils` and `testutils` packages.
Having a `utils` package is a bad practice in go.
The main reason is that the import name of the
utils sub-packages does not describe their meaning.
For example - `testutils` is more meaningful than `tests`.
